requirements.txt
asgiref==3.2.7
Django==3.0.7
django-filter==2.3.0
django-partial-date==1.3.1
djangorestframework==3.11.0
Markdown==3.2.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
